Safari-611.1.10/LayoutTests/accessibility/aria-checkbox-checked-expected.txt


X
X

This tests that ARIA checkboxes correctly handle the aria-checked attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkbox.isChecked is false
PASS checkbox.isChecked is true
PASS checkbox.isChecked is false
PASS checkbox.isChecked is true
PASS checkbox.isChecked is false
PASS checkbox.isChecked is true
PASS checkbox.isChecked is false
PASS checkbox.isChecked is true
PASS successfullyParsed is true

TEST COMPLETE
